# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindVault AI Personal Workspace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  fullName: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  fullName: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateProfileBody {
  fullName?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface Folder {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  icon?: string | null;
  isSystem: boolean;
  itemCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateFolderBody {
  name: string;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  icon?: string | null;
}

export interface UpdateFolderBody {
  name?: string;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  icon?: string | null;
}

export type ItemType = (typeof ItemType)[keyof typeof ItemType];

export const ItemType = {
  note: "note",
  file: "file",
  reminder: "reminder",
} as const;

export type ItemStatus = (typeof ItemStatus)[keyof typeof ItemStatus];

export const ItemStatus = {
  active: "active",
  archived: "archived",
  completed: "completed",
} as const;

export interface Item {
  id: number;
  userId: number;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  folderName?: string | null;
  type: ItemType;
  title: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  originalFilename?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  fileData?: string | null;
  /** @nullable */
  reminderAt?: string | null;
  status: ItemStatus;
  /** @nullable */
  aiCategory?: string | null;
  aiTags: string[];
  /** @nullable */
  aiConfidence?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateItemBodyType =
  (typeof CreateItemBodyType)[keyof typeof CreateItemBodyType];

export const CreateItemBodyType = {
  note: "note",
  file: "file",
  reminder: "reminder",
} as const;

export interface CreateItemBody {
  type: CreateItemBodyType;
  title: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  reminderAt?: string | null;
}

export type UpdateItemBodyStatus =
  (typeof UpdateItemBodyStatus)[keyof typeof UpdateItemBodyStatus];

export const UpdateItemBodyStatus = {
  active: "active",
  archived: "archived",
  completed: "completed",
} as const;

export interface UpdateItemBody {
  title?: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  reminderAt?: string | null;
  status?: UpdateItemBodyStatus;
}

export interface UploadFileBody {
  filename: string;
  mimeType: string;
  /** Base64-encoded file data */
  fileData: string;
  fileSize: number;
  /** @nullable */
  folderId?: number | null;
}

export interface ClassifyContentBody {
  content: string;
  /**
   * Optional conversation context
   * @nullable
   */
  conversationId?: number | null;
}

export type ClassifyContentResponseType =
  (typeof ClassifyContentResponseType)[keyof typeof ClassifyContentResponseType];

export const ClassifyContentResponseType = {
  note: "note",
  reminder: "reminder",
  file: "file",
  chat: "chat",
} as const;

export interface ClassifyContentResponse {
  type: ClassifyContentResponseType;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  cleanedContent?: string | null;
  /** @nullable */
  suggestedFolder?: string | null;
  tags: string[];
  confidence: number;
  /** @nullable */
  reminderAt?: string | null;
  savedItem?: Item | null;
  message: string;
}

export interface WorkspaceStats {
  totalItems: number;
  totalNotes: number;
  totalFiles: number;
  totalReminders: number;
  totalFolders: number;
  pendingReminders: number;
  totalConversations: number;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateGeminiConversationBody {
  title: string;
}

export interface SendGeminiMessageBody {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GenerateGeminiImageBody {
  prompt: string;
}

export interface GenerateGeminiImageResponse {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}

export type ListItemsParams = {
  type?: ListItemsType;
  /**
   * @nullable
   */
  folderId?: number | null;
  status?: ListItemsStatus;
  limit?: number;
  offset?: number;
};

export type ListItemsType = (typeof ListItemsType)[keyof typeof ListItemsType];

export const ListItemsType = {
  note: "note",
  file: "file",
  reminder: "reminder",
} as const;

export type ListItemsStatus =
  (typeof ListItemsStatus)[keyof typeof ListItemsStatus];

export const ListItemsStatus = {
  active: "active",
  archived: "archived",
  completed: "completed",
} as const;

export type SearchItemsParams = {
  q: string;
  type?: SearchItemsType;
  /**
   * @nullable
   */
  folderId?: number | null;
};

export type SearchItemsType =
  (typeof SearchItemsType)[keyof typeof SearchItemsType];

export const SearchItemsType = {
  note: "note",
  file: "file",
  reminder: "reminder",
} as const;

export type GetRecentItemsParams = {
  limit?: number;
};

export type GetUpcomingRemindersParams = {
  limit?: number;
};