# lib/api-zod/src/generated/types/createFolderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindVault AI Personal Workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateFolderBody {
  name: string;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  icon?: string | null;
}