# lib/api-zod/src/generated/types/folder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindVault AI Personal Workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface Folder {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  icon?: string | null;
  isSystem: boolean;
  itemCount: number;
  createdAt: Date;
  updatedAt: Date;
}